+++ Sources/RichTextKit/Extensions/CGFloat+Indent.swift
@@ -1,0 +1,19 @@
+//
+//  CGFloat+Indent.swift
+//  RichTextKit
+//
+//  Created by Daniel Saidi on 2023-10-12.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public extension CGFloat {
+    
+    static var defaultRichTextIntentStepSize: CGFloat = 30.0
+}
+
+public extension UInt {
+    
+    static var defaultRichTextIntentStepSize: UInt = 30
+}