+++ Sources/RichTextKit/_Deprecated/RichTextContext+Deprecated.swift
@@ -2,13 +2,33 @@ import SwiftUI
 
 public extension RichTextContext {
     
-    @available(*, deprecated, renamed: "decreaseFontSize(points:)")
+    @available(*, deprecated, message: "Use handle(_:) instead")
     func decrementFontSize(points: UInt = 1) {
-        stepFontSize(points: -Int(points))
+        handle(.decreaseFontSize(points: points))
     }
     
-    @available(*, deprecated, renamed: "increaseFontSize(points:)")
+    @available(*, deprecated, message: "Use handle(_:) instead")
     func incrementFontSize(points: UInt = 1) {
-        stepFontSize(points: Int(points))
+        handle(.increaseFontSize(points: points))
+    }
+    
+    @available(*, deprecated, message: "Use handle(_:) instead")
+    func increaseFontSize(points: UInt = 1) {
+        handle(.increaseFontSize(points: points))
+    }
+    
+    @available(*, deprecated, message: "Use handle(_:) instead")
+    func decreaseFontSize(points: UInt = 1) {
+        handle(.decreaseFontSize(points: points))
+    }
+    
+    @available(*, deprecated, message: "Use handle(_:) instead")
+    func increaseIndent(points: UInt = 1) {
+        handle(.increaseIndent(points: points))
+    }
+    
+    @available(*, deprecated, message: "Use handle(_:) instead")
+    func decreaseIndent(points: UInt = 1) {
+        handle(.decreaseIndent(points: points))
     }
 }